python-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.2
scipy==1.11.4
psycopg2-binary==2.9.9
pydantic==2.5.2
